+++ Assets/Plugins/GameJolt/Demo/Console/ConsoleTest.cs
@@ -1,6 +1,7 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
 using System.Linq;
+using GameJolt.API.Objects;
 
 public class ConsoleTest : MonoBehaviour
 {
@@ -43,10 +44,10 @@ public class ConsoleTest : MonoBehaviour
 
 		var user = new GameJolt.API.Objects.User(userNameField.text, userTokenField.text);
 		user.SignIn(
-			(bool signInSuccess) => {
+			signInSuccess => {
 				AddConsoleLine(string.Format("Sign In {0}.", signInSuccess ? "Successful" : "Failed"));
 			},
-			(bool userFetchSuccess) => {
+			userFetchSuccess => {
 				AddConsoleLine(string.Format("User's Information Fetch {0}.", userFetchSuccess ? "Successful" : "Failed"));
 			});
 	}
@@ -69,7 +70,7 @@ public class ConsoleTest : MonoBehaviour
 		Debug.Log("Get Users By Id. Click to see source.");
 
 		var ids = ParseIds(userIdsField.text);
-		GameJolt.API.Users.Get(ids, (GameJolt.API.Objects.User[] users) => {
+		GameJolt.API.Users.Get(ids, users => {
 			if(users != null) {
 				foreach(var user in users) {
 					AddConsoleLine(string.Format("> {0} - {1}", user.Name, user.ID));
@@ -83,7 +84,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Session Open. Click to see source.");
 
-		GameJolt.API.Sessions.Open((bool success) => {
+		GameJolt.API.Sessions.Open(success => {
 			AddConsoleLine(string.Format("Session Open {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -92,7 +93,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Session Ping Active. Click to see source.");
 
-		GameJolt.API.Sessions.Ping(GameJolt.API.SessionStatus.Active, (bool success) => {
+		GameJolt.API.Sessions.Ping(GameJolt.API.SessionStatus.Active, success => {
 			AddConsoleLine(string.Format("Session Ping Active {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -101,7 +102,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Session Ping Idle. Click to see source.");
 
-		GameJolt.API.Sessions.Ping(GameJolt.API.SessionStatus.Idle, (bool success) => {
+		GameJolt.API.Sessions.Ping(GameJolt.API.SessionStatus.Idle, success => {
 			AddConsoleLine(string.Format("Session Ping Idle {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -110,7 +111,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Session Close. Click to see source.");
 
-		GameJolt.API.Sessions.Close((bool success) => {
+		GameJolt.API.Sessions.Close(success => {
 			AddConsoleLine(string.Format("Session Close {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -119,7 +120,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Get Tables. Click to see source.");
 
-		GameJolt.API.Scores.GetTables((GameJolt.API.Objects.Table[] tables) => {
+		GameJolt.API.Scores.GetTables(tables => {
 			if (tables != null)
 			{
 				foreach (var table in tables.Reverse<GameJolt.API.Objects.Table>())
@@ -140,7 +141,7 @@ public class ConsoleTest : MonoBehaviour
 			var scoreValue = scoreValueField.text != string.Empty ? int.Parse(scoreValueField.text) : 0;
 			var tableID = tableField.text != string.Empty ? int.Parse(tableField.text) : 0;
 
-			GameJolt.API.Scores.Add(scoreValue, scoreTextField.text, guestNameField.text, tableID, "", (bool success) => {
+			GameJolt.API.Scores.Add(scoreValue, scoreTextField.text, guestNameField.text, tableID, "", success => {
 				AddConsoleLine(string.Format("Score Add (for Guest) {0}.", success ? "Successful" : "Failed"));
 			});
 		}
@@ -151,7 +152,7 @@ public class ConsoleTest : MonoBehaviour
 			var scoreValue = scoreValueField.text != string.Empty ? int.Parse(scoreValueField.text) : 0;
 			var tableID = tableField.text != string.Empty ? int.Parse(tableField.text) : 0;
 
-			GameJolt.API.Scores.Add(scoreValue, scoreTextField.text, tableID, "", (bool success) => {
+			GameJolt.API.Scores.Add(scoreValue, scoreTextField.text, tableID, "", success => {
 				AddConsoleLine(string.Format("Score Add {0}.", success ? "Successful" : "Failed"));
 			});
 		}
@@ -163,7 +164,7 @@ public class ConsoleTest : MonoBehaviour
 
 		var tableID = tableField.text != string.Empty ? int.Parse(tableField.text) : 0;
 		var limit = limitField.text != string.Empty ? int.Parse(limitField.text) : 10;
-		GameJolt.API.Scores.Get((GameJolt.API.Objects.Score[] scores) => {
+		GameJolt.API.Scores.Get(scores => {
 			if (scores != null)
 			{
 				foreach (var score in scores.Reverse<GameJolt.API.Objects.Score>())
@@ -182,7 +183,7 @@ public class ConsoleTest : MonoBehaviour
 		var scoreValue = scoreValueField.text != string.Empty ? int.Parse(scoreValueField.text) : 0;
 		var tableID = tableField.text != string.Empty ? int.Parse(tableField.text) : 0;
 
-		GameJolt.API.Scores.GetRank(scoreValue, tableID, (int rank) => {
+		GameJolt.API.Scores.GetRank(scoreValue, tableID, rank => {
 			AddConsoleLine(string.Format("Rank {0}", rank));
 		});
 	}
@@ -192,7 +193,7 @@ public class ConsoleTest : MonoBehaviour
 		Debug.Log("Unlock Trophy. Click to see source.");
 
 		var trophyID = trophyIDField.text != string.Empty ? int.Parse(trophyIDField.text) : 0;
-		GameJolt.API.Trophies.Unlock(trophyID, (bool success) => {
+		GameJolt.API.Trophies.Unlock(trophyID, success => {
 			AddConsoleLine(string.Format("Unlock Trophy {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -204,12 +205,7 @@ public class ConsoleTest : MonoBehaviour
 			Debug.Log("Get Single Trophy. Click to see source.");
 
 			var trophyID = trophyIDsField.text != string.Empty ? int.Parse(trophyIDsField.text) : 0;
-			GameJolt.API.Trophies.Get(trophyID, (GameJolt.API.Objects.Trophy trophy) => {
-				if (trophy != null)
-				{
-					AddConsoleLine(string.Format("> {0} - {1} - {2} - {3}Unlocked", trophy.Title, trophy.ID, trophy.Difficulty, trophy.Unlocked ? "" : "Not "));
-				}
-			});
+			GameJolt.API.Trophies.Get(trophyID, PrintTrophy);
 		}
 		else
 		{
@@ -217,12 +213,11 @@ public class ConsoleTest : MonoBehaviour
 			
 			var trophyIDs = ParseIds(trophyIDsField.text);
 
-			GameJolt.API.Trophies.Get(trophyIDs, (GameJolt.API.Objects.Trophy[] trophies) => {
+			GameJolt.API.Trophies.Get(trophyIDs, trophies => {
 				if (trophies != null)
 				{
-					foreach (var trophy in trophies.Reverse<GameJolt.API.Objects.Trophy>())
-					{
-						AddConsoleLine(string.Format("> {0} - {1} - {2} - {3}Unlocked", trophy.Title, trophy.ID, trophy.Difficulty, trophy.Unlocked ? "" : "Not "));
+					foreach (var trophy in trophies.Reverse()) {
+						PrintTrophy(trophy);
 					}
 					AddConsoleLine(string.Format("Found {0} trophies.", trophies.Length));
 				}
@@ -234,12 +229,11 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Get All Trophies. Click to see source.");
 
-		GameJolt.API.Trophies.Get((GameJolt.API.Objects.Trophy[] trophies) => {
+		GameJolt.API.Trophies.Get(trophies => {
 			if (trophies != null)
 			{
-				foreach (var trophy in trophies.Reverse<GameJolt.API.Objects.Trophy>())
-				{
-					AddConsoleLine(string.Format("> {0} - {1} - {2} - {3}Unlocked", trophy.Title, trophy.ID, trophy.Difficulty, trophy.Unlocked ? "" : "Not "));
+				foreach (var trophy in trophies.Reverse()) {
+					PrintTrophy(trophy);
 				}
 				AddConsoleLine(string.Format("Found {0} trophies.", trophies.Length));
 			}
@@ -250,23 +244,29 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Get Trophies by Status (Unlocked or not). Click to see source.");
 
-		GameJolt.API.Trophies.Get(unlockedTrophiesOnlyToggle.isOn, (GameJolt.API.Objects.Trophy[] trophies) => {
+		GameJolt.API.Trophies.Get(unlockedTrophiesOnlyToggle.isOn, trophies => {
 			if (trophies != null)
 			{
-				foreach (var trophy in trophies.Reverse<GameJolt.API.Objects.Trophy>())
-				{
-					AddConsoleLine(string.Format("> {0} - {1} - {2} - {3}Unlocked", trophy.Title, trophy.ID, trophy.Difficulty, trophy.Unlocked ? "" : "Not "));
+				foreach (var trophy in trophies.Reverse<GameJolt.API.Objects.Trophy>()) {
+					PrintTrophy(trophy);
 				}
 				AddConsoleLine(string.Format("Found {0} trophies.", trophies.Length));
 			}
 		});
 	}
 
+	private void PrintTrophy(Trophy trophy) {
+		if(trophy != null) {
+			AddConsoleLine(string.Format("> {0} - {1} - {2} - {3}Unlocked - {4}Secret", trophy.Title, trophy.ID,
+				trophy.Difficulty, trophy.Unlocked ? "" : "Not ", trophy.IsSecret ? "" : "Not "));
+		}
+	}
+
 	public void GetDataStoreKey()
 	{
 		Debug.Log("Get DataStore Key. Click to see source.");
 
-		GameJolt.API.DataStore.Get(keyField.text, globalToggle.isOn, (string value) => {
+		GameJolt.API.DataStore.Get(keyField.text, globalToggle.isOn, value => {
 			if (value != null)
 			{
 				AddConsoleLine(string.Format("> {0}", value));
@@ -278,7 +278,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Get DataStore Keys. Click to see source.");
 
-		GameJolt.API.DataStore.GetKeys(globalToggle.isOn, patternField.text, (string[] keys) => {
+		GameJolt.API.DataStore.GetKeys(globalToggle.isOn, patternField.text, keys => {
 			if (keys != null)
 			{
 				foreach (var key in keys)
@@ -298,7 +298,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Remove DataStore Key. Click to see source.");
 
-		GameJolt.API.DataStore.Delete(keyField.text, globalToggle.isOn, (bool success) => {
+		GameJolt.API.DataStore.Delete(keyField.text, globalToggle.isOn, success => {
 			AddConsoleLine(string.Format("Remove DataStore Key {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -307,7 +307,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Set DataStore Key. Click to see source.");
 
-		GameJolt.API.DataStore.Set(keyField.text, valueField.text, globalToggle.isOn, (bool success) => {
+		GameJolt.API.DataStore.Set(keyField.text, valueField.text, globalToggle.isOn, success => {
 			AddConsoleLine(string.Format("Set DataStore Key {0}.", success ? "Successful" : "Failed"));
 		});
 	}
@@ -327,7 +327,7 @@ public class ConsoleTest : MonoBehaviour
 
 		Debug.Log("Update DataStore Key. Click to see source.");
 
-		GameJolt.API.DataStore.Update(keyField.text, valueField.text, mode, globalToggle.isOn, (string value) => {
+		GameJolt.API.DataStore.Update(keyField.text, valueField.text, mode, globalToggle.isOn, value => {
 			if (value != null)
 			{
 				AddConsoleLine(string.Format("> {0}", value));
@@ -339,7 +339,7 @@ public class ConsoleTest : MonoBehaviour
 	{
 		Debug.Log("Get Time. Click to see source.");
 
-		GameJolt.API.Misc.GetTime((System.DateTime time) => {
+		GameJolt.API.Misc.GetTime(time => {
 			AddConsoleLine(string.Format("Server Time: {0}", time));
 		});
 	}